requirements.txt
livekit-agents[openai]>=1.0
livekit-plugins-openai>=1.0
livekit-api>=1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
